Use number of fail-lows in LMR formula instead of number of moves played (#53)
With more alpha raises, it signals that move ordering was in a suboptimal order (ideally there should only be a single alpha raise). Thus, we tone down the aggressiveness of LMR since we have less confidence in it. Note that this patch is non-functional in non PV nodes as it is impossible for a zero-window search to be exact bound.

Elo   | 1.83 +- 1.48 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.91 (-2.25, 2.89) [0.00, 3.00]
Games | N: 60050 W: 14681 L: 14364 D: 31005
Penta | [296, 7143, 14814, 7492, 280]
https://chess.swehosting.se/test/8591/

Bench 10224196